restructure spammer
